+++ src/Resource/EnvelopeProperties.php
@@ -41,4 +41,7 @@ class EnvelopeProperties extends BaseResource
 
     /** @var array<string> */
     public $timestampingAuthorities;
+
+    /** @var bool */
+    public $sendCompleted;
 }