+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeRecipient.authFailedReason`
 - Update `EnvelopesEndpoint.discard` with body like discardAt param
 - Add new `EnvelopeDocumentSignatureSheets` resource and endpoint
-- Add `Envelope.description`
+- Add `AccountSettings.useEnvelopeDescription`, `Envelope.description` and `EnvelopeTemplate.description`
 
 ## [1.10.0]
 ### Added