+++ CHANGELOG.md
@@ -15,8 +15,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Update `EnvelopesEndpoint.discard` with body like discardAt param
 - Add new `EnvelopeDocumentSignatureSheets` resource and endpoint
 - Add my envelope info endpoint
-- Add `User.autoscrollSignatureTags`
 - Add `AccountSettings.useEnvelopeDescription`, `Envelope.description` and `EnvelopeTemplate.description`
+- Add `User.autoscrollSignatureTags`
 
 ## [1.10.0]
 ### Added