+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add my envelope info endpoint
 - Add `AccountSettings.useEnvelopeDescription`, `Envelope.description` and `EnvelopeTemplate.description`
 - Add `User.autoscrollTags`
+- Add `AccountSecurity.discardedEnvelopeRetention`
 
 ## [1.10.0]
 ### Added