+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Add new fields for IdentityProvider
 
 ## [1.11.0]
 ### Added